requirements.txt
cubes==1.1
googletrans==3.1.0a0
moviepy==1.0.3
opencv-python==4.8.1.78
pydub==0.25.1
scikit-learn==1.3.1
SQLAlchemy==1.4.4
websockets==11.0.3
whisper==1.1.10
mysql-connector-python==8.0.33
mediapipe==0.10.5